refactor(web): optimizations
