+++ apps/web/src/features/users/pages/users-page.tsx
@@ -1,8 +1,11 @@
-"use client";
+import { headers } from "next/headers";
 
+import { requireCurrentUser } from "@/features/auth/lib/server";
 import { UserSearch } from "../components/user-search";
 
-export const UsersPage = () => {
+export const UsersPage = async () => {
+  await requireCurrentUser(await headers());
+
   return (
     <div className="container mx-auto px-4 py-32">
       <div className="mb-24 space-y-4 text-center">
